feat(front): handle assets & libs with webpack
