--- public/css/admin.css
@@ -1,64 +1,0 @@
-.ss-main .ss-multi-selected {
-    padding: .375rem .75rem;
-    border: 1px solid var(--bs-gray-400);
-    border-radius: 0;
-}
-.ss-main .ss-multi-selected .ss-values .ss-value {
-    background-color: var(--bs-blue);
-    font-size: 14px;
-}
-.ss-main .ss-multi-selected .ss-add {
-    margin: 12px 12px 0 5px;
-}
-
-.vich-image {
-    width: 100%;
-    max-height: 150px;
-    display: flex;
-    justify-content: center;
-}
-.vich-image-pic {
-    max-height: 150px;
-    border: 1px solid #ced4da;
-    object-fit: contain;
-}
-.vich-image #show-picture {
-    position: relative;
-}
-.vich-image #show-picture .btn {
-    position: absolute;
-    top: 5px;
-    right: 5px;
-    border: 1px solid var(--tblr-btn-border-color);
-    padding: 4px 5px 4px 7px;
-    width: 30px;
-    height: 30px;
-    line-height: 30px;
-    font-size: 16px;
-    border-radius: 0;
-}
-.vich-image #show-picture .btn:hover {
-    background-color: #555859;
-    border-color: #4b4e50;
-}
-.vich-image #edit-picture {
-    display: none;
-}
-
-.footer {
-    padding: 1rem 0;
-}
-.flashes button.btn-close {
-  filter: inherit;
-}
-
-.table-actions {
-    padding: .5rem;
-}
-.table-actions a {
-    height: 30px;
-}
-.ts-wrapper.multi .ts-control > div {
-    background: var(--tblr-blue);
-    color: var(--tblr-white);
-}
