--- public/js/admin/pictures.js
@@ -1,4 +1,0 @@
-document.getElementById("edit-btn").addEventListener("click", () => {
-    document.getElementById('edit-picture').style.display = 'block';
-    document.getElementById('show-picture').style.display = 'none';
-});
