+++ templates/admin/about_me/index.html.twig
@@ -19,8 +19,3 @@
     {{ form_end(form) }}
 </div>
 {% endblock %}
-
-{% block javascripts %}
-    {{ parent() }}
-    <script src="/js/admin/pictures.js"></script>
-{% endblock %}
