+++ templates/admin/social/edit.html.twig
@@ -17,8 +17,3 @@
     {{ form_end(form) }}
 </div>
 {% endblock %}
-
-{% block javascripts %}
-    {{ parent() }}
-    <script src="/js/admin/pictures.js"></script>
-{% endblock %}
